VRS_APP/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
auto-py-to-exe==2.36.0
autopep8==2.0.2
beautifulsoup4==4.12.2
black==23.3.0
bottle==0.12.25
bottle-websocket==0.2.9
cachetools==5.3.0
certifi==2022.12.7
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
coverage==7.2.3
cssselect==1.2.0
cycler==0.11.0
distlib==0.3.6
dnspython==2.3.0
docker==6.0.1
duckduckgo-search==2.8.5
Eel==0.16.0
exceptiongroup==1.1.1
filelock==3.11.0
flake8==6.0.0
fonttools==4.39.3
frozenlist==1.3.3
gevent==23.7.0
gevent-websocket==0.10.1
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
gTTS==2.3.1
h11==0.14.0
httplib2==0.22.0
identify==2.5.22
idna==3.4
importlib-resources==5.12.0
iniconfig==2.0.0
isort==5.12.0
jsonschema==4.17.3
kiwisolver==1.4.4
loguru==0.7.0
lxml==4.9.2
matplotlib==3.7.1
mccabe==0.7.0
mpl-interactions==0.23.0
mplcursors==0.5.2
mpldatacursor==0.7.1
mplfinance==0.12.9b7
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
nodeenv==1.7.0
numpy==1.24.2
oauthlib==3.2.2
openai==0.27.2
orjson==3.8.10
outcome==1.2.0
packaging==23.1
pandas==2.0.0
pathspec==0.11.1
Pillow==9.5.0
pinecone-client==2.2.1
platformdirs==3.2.0
playsound==1.2.2
pluggy==1.0.0
pre-commit==3.2.2
protobuf==4.22.3
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pyflakes==3.0.1
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.6
pyparsing==3.0.9
PyQt6==6.5.0
PyQt6-Qt6==6.5.0
PyQt6-sip==13.5.1
pyrsistent==0.19.3
PySide6==6.5.0
PySide6-Addons==6.5.0
PySide6-Essentials==6.5.0
PySocks==1.7.1
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-benchmark==4.0.0
pytest-cov==4.0.0
pytest-integration==0.2.3
pytest-mock==3.10.0
python-can==4.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
readability-lxml==0.8.1
redis==4.5.4
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.8.3
shiboken6==6.5.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sourcery==1.2.0
tiktoken==0.3.3
tomli==2.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
tweepy==4.13.0
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.21.0
webdriver-manager==3.8.6
websocket-client==1.5.1
whichcraft==0.6.1
wrapt==1.15.0
wsproto==1.2.0
yarl==1.8.2
zipp==3.15.0
zope.event==5.0
zope.interface==6.0
